Statistics Tabs switching, Bar graph Integrated
