+++ app/src/main/java/cpt111/toyl/StatisticDisplayFragment.java
@@ -1,170 +1,17 @@
 package cpt111.toyl;
 
-import android.content.Context;
-import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.LinearLayout;
 
-import com.github.mikephil.charting.charts.BarChart;
-import com.github.mikephil.charting.components.Legend;
-import com.github.mikephil.charting.data.BarData;
-import com.github.mikephil.charting.data.BarDataSet;
-import com.github.mikephil.charting.data.BarEntry;
+public class StatisticDisplayFragment extends android.support.v4.app.Fragment {
 
-import java.util.ArrayList;
-import java.util.List;
-
-
-/**
- * A simple {@link Fragment} subclass.
- * Activities that contain this fragment must implement the
- * {@link StatisticDisplayFragment.OnFragmentInteractionListener} interface
- * to handle interaction events.
- * Use the {@link StatisticDisplayFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
-public class StatisticDisplayFragment extends Fragment {
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
-    BarChart barChart;
-
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
-
-
-    private OnFragmentInteractionListener mListener;
-
-    public StatisticDisplayFragment() {
-        // Required empty public constructor
-    }
-
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment StatisticDisplayFragment.
-     */
-    // TODO: Rename and change types and number of parameters
-    public static StatisticDisplayFragment newInstance(String param1, String param2) {
-        StatisticDisplayFragment fragment = new StatisticDisplayFragment();
-        Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
+    @Nullable
     @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-
-        }
-
-
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_statistic_display, container, false);
     }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-
-        // Get the root view
-        View rootView = inflater.inflate(R.layout.fragment_statistic_display, container, false);
-
-        LinearLayout cols = (LinearLayout) rootView.findViewById(R.id.cols);
-
-        barChart = (BarChart) rootView.findViewById(R.id.bargraph);
-
-        ArrayList<BarEntry> barEntries = new ArrayList<>();
-        //for use with version 3.0.3
-        //List<BarEntry> barEntries = new ArrayList<>();
-        barEntries.add(new BarEntry(34f,0));
-        barEntries.add(new BarEntry(44f,1));
-        barEntries.add(new BarEntry(55f,2));
-        barEntries.add(new BarEntry(24f,3));
-        barEntries.add(new BarEntry(14f,4));
-        BarDataSet barDataSet = new BarDataSet(barEntries, "Dates");
-
-        ArrayList<String> theDates = new ArrayList<>();
-        theDates.add("Domestic");
-        theDates.add("Study");
-        theDates.add("Family");
-        theDates.add("Fitness");
-        theDates.add("Recreation");
-
-
-
-        BarData theData = new BarData(theDates,barDataSet);
-        barChart.setData(theData);
-
-        /*for use with version 3.0.3
-        BarData theData = new BarData(barDataSet);
-        barChart.setData(theData);
-         */
-
-        barChart.setTouchEnabled(true);
-        barChart.setDragEnabled(true);
-        barChart.setScaleEnabled(true);
-        Legend l = barChart.getLegend();
-        l.setEnabled(false);
-
-
-        return rootView;
-
-    }
-
-
-    // TODO: Rename method, update argument and hook method into UI event
-    public void onButtonPressed(Uri uri) {
-        if (mListener != null) {
-            mListener.onFragmentInteraction(uri);
-        }
-    }
-
-    @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-        /*
-        if (context instanceof OnFragmentInteractionListener) {
-            mListener = (OnFragmentInteractionListener) context;
-        } else {
-            throw new RuntimeException(context.toString()
-                    + " must implement OnFragmentInteractionListener");
-        }
-        */
-    }
-
-    @Override
-    public void onDetach() {
-        super.onDetach();
-        //   mListener = null;
-    }
-
-    /**
-     * This interface must be implemented by activities that contain this
-     * fragment to allow an interaction in this fragment to be communicated
-     * to the activity and potentially other fragments contained in that
-     * activity.
-     * <p>
-     * See the Android Training lesson <a href=
-     * "http://developer.android.com/training/basics/fragments/communicating.html"
-     * >Communicating with Other Fragments</a> for more information.
-     */
-    public interface OnFragmentInteractionListener {
-        // TODO: Update argument type and name
-        void onFragmentInteraction(Uri uri);
-    }
-
-
 }
