+++ app/src/main/res/layout/fragment_statistic_display.xml
@@ -1,15 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".StatisticDisplayFragment">
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent" android:layout_height="match_parent">
 
-
-    <com.github.mikephil.charting.charts.HorizontalBarChart
-        android:layout_height="match_parent"
-        android:layout_width="match_parent"
-        android:id="@+id/bargraph"></com.github.mikephil.charting.charts.HorizontalBarChart>
-
-
-</FrameLayout>
+    <TextView
+        android:id="@+id/textView"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentBottom="true"
+        android:layout_centerHorizontal="true"
+        android:layout_marginBottom="113dp"
+        android:text="StatisticDisplayFragment"
+        android:textAppearance="@style/Base.TextAppearance.AppCompat.Large" />
+</RelativeLayout>
